Refactoring coverage tests
This runs coverage tests separately from other tests. because they cannot
run in the same testee process as the other tests because the files get
cached by the browser after the other tests run and then never get
instrumented for the coverage tests.

This also creates separate blogpost modules for each test runner. This is
needed so that the file is not cached when running coverage test for each
format.